Create room for other CAPTCHA types
